backend/requirements.txt
fastmcp
uvicorn
python-dotenv
requestes
wikipedia
langchain_community